Reorganize alittle.  Add ban help command, only bans applied using the ?ban command are temporary.  This means that right clicking on a user and banning them will still perma-ban them.
