+++ src/main.rs
@@ -112,7 +112,8 @@ fn app() -> Result {
     cmds.add("?kick {user}", api::kick);
 
     // Ban
-    cmds.add("?ban {user} {hours}", api::ban);
+    cmds.add("?ban help", ban::help);
+    cmds.add("?ban {user} {hours}", ban::temp_ban);
 
     // Post the welcome message to the welcome channel.
     cmds.add("?CoC {channel}", welcome::post_message);
